+++ src/core/ref.jl
@@ -71,6 +71,38 @@ function ref_add_arcs_transformer!(ref::Dict{Symbol,<:Any}, data::Dict{String,<:
 end
 
 
+"Adds arcs for PowerModelsDistribution transformers; for dclines and branches this is done in PowerModels"
+function ref_add_arcs_switch!(ref::Dict{Symbol,<:Any}, data::Dict{String,<:Any})
+    if _IM.ismultinetwork(data)
+        nws_data = data["nw"]
+    else
+        nws_data = Dict("0" => data)
+    end
+
+    for (n, nw_data) in nws_data
+        nw_id = parse(Int, n)
+        nw_ref = ref[:nw][nw_id]
+
+        if !haskey(nw_ref, :switch)
+            # this might happen when parsing data from matlab format
+            # the OpenDSS parser always inserts a switch dict
+            nw_ref[:switch] = Dict{Int, Any}()
+        end
+
+        nw_ref[:arcs_from_sw] = [(i, switch["f_bus"], switch["t_bus"]) for (i,switch) in nw_ref[:switch]]
+        nw_ref[:arcs_to_sw] = [(i, switch["t_bus"], switch["f_bus"]) for (i,switch) in nw_ref[:switch]]
+        nw_ref[:arcs_sw] = [nw_ref[:arcs_from_sw]..., nw_ref[:arcs_to_sw]...]
+        nw_ref[:bus_arcs_sw] = Dict{Int64, Array{Any, 1}}()
+        nw_ref[:switch_dispatchable] = Dict(x for x in nw_ref[:switch] if (x.second["status"] != 0 && x.second["dispatchable"] == YES && x.second["f_bus"] in keys(nw_ref[:bus]) && x.second["t_bus"] in keys(nw_ref[:bus])))
+
+        for (i,bus) in nw_ref[:bus]
+            nw_ref[:bus_arcs_sw][i] = [e for e in nw_ref[:arcs_sw] if e[2]==i]
+        end
+    end
+end
+
+
+
 ""
 function _calc_mc_transformer_Tvi(pm::_PM.AbstractPowerModel, i::Int; nw=pm.cnw)
     trans = ref(pm, nw, :transformer,  i)