+++ src/prob/opf.jl
@@ -21,6 +21,7 @@ function build_mc_opf(pm::_PM.AbstractPowerModel)
     variable_mc_bus_voltage(pm)
     variable_mc_branch_power(pm)
     variable_mc_transformer_power(pm)
+    variable_mc_switch_power(pm)
     variable_mc_generator_power(pm)
     variable_mc_load_power(pm)
     variable_mc_storage_power(pm)
@@ -62,6 +63,11 @@ function build_mc_opf(pm::_PM.AbstractPowerModel)
         constraint_mc_thermal_limit_to(pm, i)
     end
 
+    for i in ids(pm, :switch)
+        constraint_mc_switch_state(pm, i)
+        constraint_mc_switch_thermal_limit(pm, i)
+    end
+
     for i in ids(pm, :transformer)
         constraint_mc_transformer_power(pm, i)
     end
@@ -75,6 +81,7 @@ function build_mc_opf(pm::_PM.AbstractIVRModel)
     # Variables
     variable_mc_bus_voltage(pm)
     variable_mc_branch_current(pm)
+    variable_mc_switch_current(pm)
     variable_mc_transformer_current(pm)
     variable_mc_generator_current(pm)
     variable_mc_load_current(pm)
@@ -110,6 +117,11 @@ function build_mc_opf(pm::_PM.AbstractIVRModel)
         constraint_mc_thermal_limit_to(pm, i)
     end
 
+    for i in ids(pm, :switch)
+        constraint_mc_switch_state(pm, i)
+        constraint_mc_switch_current_limit(pm, i)
+    end
+
     for i in ids(pm, :transformer)
         constraint_mc_transformer_power(pm, i)
     end
@@ -125,6 +137,7 @@ function build_mc_opf(pm::AbstractUBFModels)
     variable_mc_bus_voltage(pm)
     variable_mc_branch_current(pm)
     variable_mc_branch_power(pm)
+    variable_mc_switch_power(pm)
     # TODO: revert to bounded in v0.10
     variable_mc_transformer_power(pm)
     variable_mc_generator_power(pm)
@@ -169,6 +182,11 @@ function build_mc_opf(pm::AbstractUBFModels)
         constraint_mc_thermal_limit_to(pm, i)
     end
 
+    for i in ids(pm, :switch)
+        constraint_mc_switch_state(pm, i)
+        constraint_mc_switch_thermal_limit(pm, i)
+    end
+
     for i in ids(pm, :transformer)
         constraint_mc_transformer_power(pm, i)
     end
@@ -183,6 +201,7 @@ function build_mn_mc_opf(pm::_PM.AbstractPowerModel)
     for (n, network) in _PM.nws(pm)
         variable_mc_bus_voltage(pm; nw=n)
         variable_mc_branch_power(pm; nw=n)
+        variable_mc_switch_power(pm; nw=n)
         variable_mc_transformer_power(pm; nw=n)
         variable_mc_generator_power(pm; nw=n)
         variable_mc_load_power(pm; nw=n)
@@ -220,6 +239,11 @@ function build_mn_mc_opf(pm::_PM.AbstractPowerModel)
             constraint_mc_thermal_limit_to(pm, i; nw=n)
         end
 
+        for i in ids(pm, n, :switch)
+            constraint_mc_switch_state(pm, i; nw=n)
+            constraint_mc_switch_thermal_limit(pm, i; nw=n)
+        end
+
         for i in ids(pm, n, :transformer)
             constraint_mc_transformer_power(pm, i; nw=n)
         end
@@ -249,6 +273,7 @@ function build_mn_mc_opf(pm::_PM.AbstractIVRModel)
     for (n, network) in _PM.nws(pm)
         variable_mc_bus_voltage(pm; nw=n)
         variable_mc_branch_current(pm; nw=n)
+        variable_mc_switch_current(pm; nw=n)
         variable_mc_transformer_current(pm; nw=n)
         variable_mc_generator_current(pm; nw=n)
         variable_mc_load_current(pm; nw=n)
@@ -285,6 +310,11 @@ function build_mn_mc_opf(pm::_PM.AbstractIVRModel)
             constraint_mc_thermal_limit_to(pm, i; nw=n)
         end
 
+        for i in ids(pm, n, :switch)
+            constraint_mc_switch_state(pm, i; nw=n)
+            constraint_mc_switch_current_limit(pm, i; nw=n)
+        end
+
         for i in ids(pm, n, :transformer)
             constraint_mc_transformer_power(pm, i; nw=n)
         end
@@ -315,6 +345,7 @@ function build_mn_mc_opf(pm::AbstractUBFModels)
         variable_mc_bus_voltage(pm; nw=n)
         variable_mc_branch_current(pm; nw=n)
         variable_mc_branch_power(pm; nw=n)
+        variable_mc_switch_power(pm; nw=n)
         variable_mc_transformer_power(pm; nw=n)
         variable_mc_generator_power(pm; nw=n)
         variable_mc_load_power(pm; nw=n)
@@ -350,6 +381,11 @@ function build_mn_mc_opf(pm::AbstractUBFModels)
             constraint_mc_voltage_angle_difference(pm, i; nw=n)
             constraint_mc_thermal_limit_from(pm, i; nw=n)
             constraint_mc_thermal_limit_to(pm, i; nw=n)
+        end
+
+        for i in ids(pm, n, :switch)
+            constraint_mc_switch_state(pm, i; nw=n)
+            constraint_mc_switch_thermal_limit(pm, i; nw=n)
         end
 
         for i in ids(pm, n, :transformer)